systems/outlook_kpi_scraper/requirements.txt
pywin32
python-dotenv
pyyaml
requests
beautifulsoup4
google-api-python-client
google-auth
google-auth-oauthlib
pandas
openai
